Make all models train with complete vocabulary
In order to apply the procustes rotation between embeddings, the
embeddings need to have the same overall vocabulary and the same
ordering of that vocabulary.  To make this happen we need to tell the
training algorithm about the complete vocabulary, even though the input
data will only contain a portion of it.

However, because we will still care about words that are added or
retired in any interval, continue to collect the per window vocabulary
as well.

Signed-off-by: Eric B Munson <ebm@cs.toronto.edu>